requirements.txt
# CS2 Heightmap Generator Requirements
# Python 3.8+

# Core image processing
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0

# 3D visualization (for 3D preview feature)
matplotlib>=3.5.0

# Procedural noise generation
perlin-noise>=1.12  # Fallback only (slow: 60-120s per generation)
opensimplex>=0.4.5  # Fallback only (slow: 90-150s per generation)

# CRITICAL: FastNoiseLite provides 60-140x speedup (REQUIRED for good performance)
# Without this, terrain generation takes 60-120s instead of <1s
pyfastnoiselite>=0.0.6

# Progress tracking
tqdm>=4.65.0

# Optional: GUI support (uncomment if needed)
# PyQt5>=5.15.0

# Optional: Real-world elevation data (uncomment if needed)
# elevation>=1.1.3
# srtm.py>=0.3.7
